+++ src/renderer.h
@@ -29,8 +29,8 @@ class Renderer {
         void drawRect(int16_t x, int16_t y, uint8_t w, uint8_t h, uint8_t color = WHITE);
         void drawOverwrite(int16_t x, int16_t y, const uint8_t* bitmap, uint8_t frame, MirrorMode mirror = 0, DrawMode drawMode = Normal);
         void drawPlusMask(int16_t x, int16_t y, const uint8_t* bitmap, uint8_t frame, MirrorMode mirror = 0, DrawMode drawMode = Normal);
-        void drawString(int8_t x, int8_t y, const uint8_t* str);
-        void drawNumber(int8_t x, int8_t y, uint16_t number);
+        void drawString(int16_t x, int16_t y, const uint8_t* str);
+        void drawNumber(int16_t x, int16_t y, uint16_t number);
 };
 
 #endif
